feat: add support for custom serializer
