+++ src/PostHog/Library/HttpClientExtensions.cs
@@ -13,11 +13,12 @@ internal static class HttpClientExtensions
 {
     /// <summary>
     /// Sends a POST request to the specified Uri containing the value serialized as JSON in the request body.
-    /// Returns the response body deserialized as <typeparamref name="TBody"/>.
+    /// Returns the response body deserialized as <typeparamref name="TBody"/> using the provided JSON serializer.
     /// </summary>
     /// <param name="httpClient">The client used to send the request.</param>
     /// <param name="requestUri">The Uri the request is sent to.</param>
     /// <param name="content">The value to serialize.</param>
+    /// <param name="jsonSerializer">The JSON serializer wrapper to use for deserialization.</param>
     /// <param name="cancellationToken">The cancellation token that can be used to cancel the operation.</param>
     /// <typeparam name="TBody">The type of the response body to deserialize to.</typeparam>
     /// <returns>The task representing the asynchronous operation.</returns>
@@ -25,18 +26,19 @@ internal static class HttpClientExtensions
         this HttpClient httpClient,
         Uri requestUri,
         object content,
+        JsonSerializerWrapper jsonSerializer,
         CancellationToken cancellationToken)
     {
         var response = await httpClient.PostAsJsonAsync(
             requestUri,
             content,
-            JsonSerializerHelper.Options,
+            JsonSerializerWrapper.Options,
             cancellationToken);
 
-        await response.EnsureSuccessfulApiCall(cancellationToken);
+        await response.EnsureSuccessfulApiCall(jsonSerializer, cancellationToken);
 
         var result = await response.Content.ReadAsStreamAsync(cancellationToken);
-        return await JsonSerializerHelper.DeserializeFromCamelCaseJsonAsync<TBody>(
+        return await jsonSerializer.DeserializeFromCamelCaseJsonAsync<TBody>(
             result,
             cancellationToken: cancellationToken);
     }
@@ -73,6 +75,46 @@ internal static class HttpClientExtensions
 #pragma warning disable CA2016
                 var result = await response.Content.ReadFromJsonAsync<ApiErrorResult>(
                     cancellationToken: cancellationToken);
+                return (result, null);
+            }
+            catch (JsonException e)
+            {
+                return (null, e);
+            }
+        }
+    }
+
+    public static async Task EnsureSuccessfulApiCall(
+        this HttpResponseMessage response,
+        JsonSerializerWrapper jsonSerializer,
+        CancellationToken cancellationToken)
+    {
+        if (response.IsSuccessStatusCode)
+        {
+            return;
+        }
+
+        if (response.StatusCode == HttpStatusCode.NotFound)
+        {
+            // Allow 404 exception to propagate up.
+            response.EnsureSuccessStatusCode();
+        }
+
+        var (error, exception) = await ReadApiErrorResultAsync();
+
+        throw response.StatusCode switch
+        {
+            HttpStatusCode.Unauthorized => new UnauthorizedAccessException(
+                error?.Detail ?? "Unauthorized. Could not deserialize the response for more info.", exception),
+            _ => new ApiException(error, response.StatusCode, exception)
+        };
+
+        async Task<(ApiErrorResult?, Exception?)> ReadApiErrorResultAsync()
+        {
+            try
+            {
+                var stream = await response.Content.ReadAsStreamAsync(cancellationToken);
+                var result = await jsonSerializer.DeserializeFromCamelCaseJsonAsync<ApiErrorResult>(stream, cancellationToken);
                 return (result, null);
             }
             catch (JsonException e)
